+++ enoslib/infra/enos_g5k/constants.py
@@ -37,7 +37,14 @@ PROD = "prod"
 PROD_VLAN_ID = "DEFAULT"
 
 NETWORK_TYPES = [PROD] + KAVLAN_TYPE + SUBNET_TYPES
-JOB_TYPES = ["deploy", "allow_classic_ssh", "exotic"]
+JOB_TYPES_REGEXP = [
+    "^deploy$",
+    "^allow_classic_ssh$",
+    "^exotic$",
+    "^container$",
+    "^noop$",
+    r"^inner=\d+$",
+]
 QUEUE_TYPES = ["default", "production", "testing", "besteffort"]